Brought over builder-related schemas to builder.py
